Renamed files and deleted redundant ones
